fix: singup.js, main.js
Move main.js code to signup.js
