+++ index.html
@@ -342,6 +342,9 @@
           />
         </p>
         <input type="submit" value="전송" class="signup-button" />
+        <div class="login-error-box signup-error-box">
+          회원가입 정보 들을 정확히 입력해주세요.
+        </div>
       </form>
       <button class="close-signup close">
         <i class="fa-solid fa-circle-xmark"></i>
