+++ js/signup.js
@@ -24,8 +24,10 @@ import {
   signupPwBox,
   pwErrorMsg,
   signupRepwBox,
+  exptext,
+  signupErrorBox,
 } from "./store.js";
-import { editUser } from "./requests.js";
+import { editUser, authLogin } from "./requests.js";
 const state = {
   email: "",
   password: "",
@@ -81,11 +83,13 @@ export async function createSubmitEvent(event) {
   ) {
     await signup(state.email, state.password, state.displayName);
     location.reload();
+  } else {
+    showErrorBox(signupErrorBox);
   }
 }
 
 // 유효성 검사 스타일
-export async function validationStyle(errormsg, type, element, color) {
+async function validationStyle(errormsg, type, element, color) {
   switch (type) {
     case "add":
       errormsg.classList.add("show");
@@ -100,6 +104,45 @@ export async function validationStyle(errormsg, type, element, color) {
   }
 }
 
+// 로그인 시 유효성 검사
+loginId.addEventListener("focusout", () => {
+  if (loginId.value && !exptext.test(loginId.value)) {
+    validationStyle(idErrorMsg, "add", idboxEl, "#ed234b");
+  }
+});
+loginId.addEventListener("focusin", () => {
+  validationStyle(idErrorMsg, "remove", idboxEl, "#999");
+});
+
+// 회원가입 유효성 검사
+// 이메일
+emailInputEl.addEventListener("focusout", () => {
+  if (emailInputEl.value && !exptext.test(emailInputEl.value)) {
+    validationStyle(emailErrorMsg, "add", singupEmailBox, "#ed234b");
+  }
+});
+emailInputEl.addEventListener("focusin", () => {
+  validationStyle(emailErrorMsg, "remove", singupEmailBox, "#333");
+});
+// 비밀번호 8자리 이상
+passwordInputEl.addEventListener("focusout", () => {
+  if (passwordInputEl.value && passwordInputEl.value.length < 8) {
+    validationStyle(pwLengthMsg, "add", signupPwBox, "#ed234b");
+  }
+});
+passwordInputEl.addEventListener("focusin", () => {
+  validationStyle(pwLengthMsg, "remove", signupPwBox, "#333");
+});
+// 비밀번호 확인
+passwordcheckEl.addEventListener("focusout", () => {
+  if (passwordInputEl && passwordInputEl.value !== passwordcheckEl.value) {
+    validationStyle(pwErrorMsg, "add", signupRepwBox, "#ed234b");
+  }
+});
+passwordcheckEl.addEventListener("focusin", () => {
+  validationStyle(pwErrorMsg, "remove", signupRepwBox, "#333");
+});
+
 // 로그인 이벤트
 export async function createLoginEvent(event) {
   event.preventDefault();
@@ -109,28 +152,43 @@ export async function createLoginEvent(event) {
 }
 
 // 로그인 실패 시 Error Box
-export function showErrorBox() {
-  loginErrorBox.classList.add("show");
+export function showErrorBox(errorbox) {
+  errorbox.classList.add("show");
   setTimeout(() => {
-    loginErrorBox.classList.remove("show");
+    errorbox.classList.remove("show");
   }, 1500);
+}
+
+// token이 없을 때 회원정보를 클릭하면 로그인을 하라고 모달창
+export async function userinfoClick() {
+  const token = localStorage.getItem("token");
+  if (token) {
+    window.location = "#/user";
+  } else {
+    userModalContent.innerHTML = `로그인을 해주세요.`;
+    userModal.classList.add("show");
+  }
 }
 
 // 회원정보 페이지 비밀번호 변경
 export async function pwchange(event) {
   event.preventDefault();
-  if (userInfoNewPw.value && userInfoNewPw.value.length < 8) {
+  if (!userInfoPw.value || !userInfoNewPw.value) {
+    userModalContent.innerHTML = `비밀번호를 입력해주세요.`;
+    userModal.classList.add("show");
+    return;
+  }
+  if (userInfoNewPw.value.length < 8) {
     userModalContent.innerHTML = `비밀번호를 8자리 이상 입력해주세요.`;
     userModal.classList.add("show");
     return;
   }
-  if (userInfoPw.value && userInfoNewPw.value)
-    await editUser(
-      "비밀번호",
-      userInfoName.value,
-      userInfoPw.value,
-      userInfoNewPw.value
-    );
+  await editUser(
+    "비밀번호",
+    userInfoName.value,
+    userInfoPw.value,
+    userInfoNewPw.value
+  );
 }
 
 // 만료 시간과 함께 데이터를 저장
@@ -166,4 +224,17 @@ export function getItemWithExpireTime(keyName) {
   }
   // 만료기간이 남아있는 경우, value 값 리턴
   return obj.value;
+}
+
+// 로컬에 로그인 데이터 있는지 확인.
+export async function autoLogin() {
+  const token = localStorage.getItem("token");
+  if (token) {
+    loginBtnEl.textContent = "로그아웃";
+    await authLogin();
+  } else {
+    loginBtnEl.textContent = "로그인/가입";
+  }
+  // 만료시간 체크는 계속
+  getItemWithExpireTime("token");
 }
